fix: move locking logic to Main
