+++ p2p_wallet/Services/AuthenticationHandler.swift
@@ -13,6 +13,7 @@ protocol AuthenticationHandlerType {
     func authenticate(presentationStyle: AuthenticationPresentationStyle?)
     func pauseAuthentication(_ isPaused: Bool)
     var authenticationStatusDriver: Driver<AuthenticationPresentationStyle?> { get }
+    var isLockedDriver: Driver<Bool> { get }
 }
 
 final class AuthenticationHandler: AuthenticationHandlerType {
@@ -26,6 +27,7 @@ final class AuthenticationHandler: AuthenticationHandlerType {
     // MARK: - Subjects
 
     private let authenticationStatusSubject = BehaviorRelay<AuthenticationPresentationStyle?>(value: nil)
+    private let isLockedSubject = BehaviorRelay<Bool>(value: false)
 
     init() {
         bind()
@@ -46,13 +48,24 @@ final class AuthenticationHandler: AuthenticationHandlerType {
     }
 
     private func observeAppNotifications() {
+        UIApplication.shared.rx
+            .applicationWillResignActive
+            .subscribe(onNext: { [weak self] _ in
+                self?.isLockedSubject.accept(true)
+            })
+            .disposed(by: disposeBag)
+
         UIApplication.shared.rx.applicationDidBecomeActive
             .subscribe(onNext: { [weak self] _ in
                 guard let self = self else { return }
                 guard Int(Date().timeIntervalSince1970) >= self.lastAuthenticationTimeStamp + self
                     .timeRequiredForAuthentication
-                else { return }
+                else {
+                    self.isLockedSubject.accept(false)
+                    return
+                }
                 self.authenticate(presentationStyle: .login())
+                self.isLockedSubject.accept(false)
             })
             .disposed(by: disposeBag)
     }
@@ -69,6 +82,7 @@ final class AuthenticationHandler: AuthenticationHandlerType {
         // accept current if nil
         if current == nil {
             authenticationStatusSubject.accept(nil)
+            isLockedSubject.accept(false)
             return
         }
 
@@ -81,6 +95,7 @@ final class AuthenticationHandler: AuthenticationHandlerType {
         // force a dissmision if not
         else {
             authenticationStatusSubject.accept(nil)
+            isLockedSubject.accept(false)
             return
         }
     }
@@ -91,6 +106,10 @@ final class AuthenticationHandler: AuthenticationHandlerType {
 
     var authenticationStatusDriver: Driver<AuthenticationPresentationStyle?> {
         authenticationStatusSubject.asDriver()
+    }
+
+    var isLockedDriver: Driver<Bool> {
+        isLockedSubject.asDriver()
     }
 
     // MARK: - Helpers
